# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Example of rest API written in Golang, used for learning purposes.",
        "title": "Example REST API",
        "contact": {
            "name": "Matěj Uchytil",
            "url": "https://github.com/uchytma/go-example-rest-api"
        },
        "version": "1.0"
    },
    "paths": {
        "/math/greatest-common-divisor": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/math"
                ],
                "summary": "Return greater common divisor (GCD) of specified numbers.",
                "operationId": "math/greatest-common-divisor",
                "parameters": [
                    {
                        "description": "All numbers must be greater than zero.",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
        "/math/least-common-multiple": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/math"
                ],
                "summary": "Return least common multiple (LCM) of specified numbers.",
                "operationId": "math/least-common-multiple",
                "parameters": [
                    {
                        "description": "All numbers must be greater than zero.",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    }
                }
            }
        }
    }
}